rename productredesign to product
