+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/ui/ProductFormState.kt
@@ -1,4 +1,4 @@
-package com.maksimowiczm.foodyou.feature.productredesign.ui
+package com.maksimowiczm.foodyou.feature.product.ui
 
 import androidx.compose.runtime.Composable
 import com.maksimowiczm.foodyou.core.input.Input
