+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/ui/UpdateProductScreen.kt
@@ -1,8 +1,8 @@
-package com.maksimowiczm.foodyou.feature.productredesign
+package com.maksimowiczm.foodyou.feature.product.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.maksimowiczm.foodyou.feature.productredesign.ui.update.UpdateProductApp
+import com.maksimowiczm.foodyou.feature.product.ui.update.UpdateProductApp
 
 @Composable
 fun UpdateProductScreen(
