+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/ui/create/CreateProductViewModel.kt
@@ -1,17 +1,18 @@
-package com.maksimowiczm.foodyou.feature.productredesign.ui.update
+package com.maksimowiczm.foodyou.feature.product.ui.create
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.maksimowiczm.foodyou.core.domain.model.Product
+import com.maksimowiczm.foodyou.core.domain.model.openfoodfacts.OpenFoodFactsProduct
+import com.maksimowiczm.foodyou.core.domain.source.OpenFoodFactsRemoteDataSource
 import com.maksimowiczm.foodyou.core.input.Form
 import com.maksimowiczm.foodyou.core.input.Input
 import com.maksimowiczm.foodyou.core.input.ValidationStrategy
 import com.maksimowiczm.foodyou.core.input.dsl.input
 import com.maksimowiczm.foodyou.core.ui.res.formatClipZeros
-import com.maksimowiczm.foodyou.feature.productredesign.data.ProductRepository
-import com.maksimowiczm.foodyou.feature.productredesign.ui.ProductFormFieldError
-import com.maksimowiczm.foodyou.feature.productredesign.ui.ProductFormRules
-import com.maksimowiczm.foodyou.feature.productredesign.ui.ProductFormState
+import com.maksimowiczm.foodyou.feature.product.data.ProductRepository
+import com.maksimowiczm.foodyou.feature.product.ui.ProductFormFieldError
+import com.maksimowiczm.foodyou.feature.product.ui.ProductFormRules
+import com.maksimowiczm.foodyou.feature.product.ui.ProductFormState
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -19,40 +20,18 @@ import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
-internal class UpdateProductViewModel(
-    private val productId: Long,
-    private val productRepository: ProductRepository
+internal class CreateProductViewModel(
+    private val productRepository: ProductRepository,
+    private val openFoodFactsRemoteDataSource: OpenFoodFactsRemoteDataSource
 ) : ViewModel() {
-    private var product: Product? = null
-    private val _formState = MutableStateFlow<ProductFormState?>(null)
+    private val _formState = MutableStateFlow<ProductFormState>(ProductFormState())
     val formState = _formState.asStateFlow()
 
-    init {
-        viewModelScope.launch {
-            val product = productRepository.getProductById(productId) ?: return@launch
+    private val _isDownloading = MutableStateFlow(false)
+    val isDownloading = _isDownloading.asStateFlow()
 
-            this@UpdateProductViewModel.product = product
-
-            _formState.value = ProductFormState(
-                name = input(product.name),
-                brand = input(product.brand ?: ""),
-                barcode = input(product.barcode ?: ""),
-                proteins = input(product.nutrients.proteins.value.formatClipZeros()),
-                carbohydrates = input(product.nutrients.carbohydrates.value.formatClipZeros()),
-                fats = input(product.nutrients.fats.value.formatClipZeros()),
-                sugars = input(product.nutrients.sugars.value?.formatClipZeros() ?: ""),
-                saturatedFats = input(
-                    product.nutrients.saturatedFats.value?.formatClipZeros() ?: ""
-                ),
-                salt = input(product.nutrients.salt.value?.formatClipZeros() ?: ""),
-                sodium = input(product.nutrients.sodium.value?.formatClipZeros() ?: ""),
-                fiber = input(product.nutrients.fiber.value?.formatClipZeros() ?: ""),
-                packageWeight = input(product.packageWeight?.weight?.formatClipZeros() ?: ""),
-                servingWeight = input(product.servingWeight?.weight?.formatClipZeros() ?: ""),
-                isModified = false
-            )
-        }
-    }
+    private val _eventBus = Channel<ProductFormEvent>()
+    val eventBus = _eventBus.receiveAsFlow()
 
     private val nameForm = Form<ProductFormFieldError>(
         ValidationStrategy.LazyEval,
@@ -61,10 +40,10 @@ internal class UpdateProductViewModel(
 
     fun onNameChange(name: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 name = nameForm.validate(name)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -73,10 +52,10 @@ internal class UpdateProductViewModel(
     private val brandForm = Form<ProductFormFieldError>(ValidationStrategy.LazyEval)
     fun onBrandChange(brand: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 brand = brandForm.validate(brand)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -85,10 +64,10 @@ internal class UpdateProductViewModel(
     private val barcodeForm = Form<ProductFormFieldError>(ValidationStrategy.LazyEval)
     fun onBarcodeChange(barcode: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 barcode = barcodeForm.validate(barcode)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -103,10 +82,10 @@ internal class UpdateProductViewModel(
 
     fun onProteinsChange(proteins: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 proteins = proteinsForm.validate(proteins)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -121,10 +100,10 @@ internal class UpdateProductViewModel(
 
     fun onCarbohydratesChange(carbohydrates: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 carbohydrates = carbohydratesForm.validate(carbohydrates)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -139,10 +118,10 @@ internal class UpdateProductViewModel(
 
     fun onFatsChange(fats: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 fats = fatsForm.validate(fats)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -156,10 +135,10 @@ internal class UpdateProductViewModel(
 
     fun onSugarsChange(sugars: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 sugars = sugarsForm.validate(sugars)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -173,10 +152,10 @@ internal class UpdateProductViewModel(
 
     fun onSaturatedFatsChange(saturatedFats: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 saturatedFats = saturatedFatsForm.validate(saturatedFats)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -190,10 +169,10 @@ internal class UpdateProductViewModel(
 
     fun onSaltChange(salt: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 salt = saltForm.validate(salt)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -207,10 +186,10 @@ internal class UpdateProductViewModel(
 
     fun onSodiumChange(sodium: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 sodium = sodiumForm.validate(sodium)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -224,10 +203,10 @@ internal class UpdateProductViewModel(
 
     fun onFiberChange(fiber: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 fiber = fiberForm.validate(fiber)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -241,10 +220,10 @@ internal class UpdateProductViewModel(
 
     fun onPackageWeightChange(packageWeight: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 packageWeight = packageWeightForm.validate(packageWeight)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
@@ -258,39 +237,34 @@ internal class UpdateProductViewModel(
 
     fun onServingWeightChange(servingWeight: String) {
         _formState.update {
-            val newState = it?.copy(
+            val newState = it.copy(
                 servingWeight = servingWeightForm.validate(servingWeight)
             )
-            newState?.copy(
+            newState.copy(
                 isModified = isModified(newState)
             )
         }
     }
 
-    private fun Input<ProductFormFieldError>.float(): Float? = value.toFloatOrNull()
-
     private fun isModified(formState: ProductFormState) = when {
-        formState.name.value != product?.name -> true
-        formState.brand.value != product?.brand -> true
-        formState.barcode.value != product?.barcode -> true
-        formState.proteins.float() != product?.nutrients?.proteins?.value -> true
-        formState.carbohydrates.float() != product?.nutrients?.carbohydrates?.value -> true
-        formState.fats.float() != product?.nutrients?.fats?.value -> true
-        formState.sugars.float() != product?.nutrients?.sugars?.value -> true
-        formState.saturatedFats.float() != product?.nutrients?.saturatedFats?.value -> true
-        formState.salt.float() != product?.nutrients?.salt?.value -> true
-        formState.sodium.float() != product?.nutrients?.sodium?.value -> true
-        formState.fiber.float() != product?.nutrients?.fiber?.value -> true
-        formState.packageWeight.float() != product?.packageWeight?.weight -> true
-        formState.servingWeight.float() != product?.servingWeight?.weight -> true
+        formState.name !is Input.Empty -> true
+        formState.brand !is Input.Empty -> true
+        formState.barcode !is Input.Empty -> true
+        formState.proteins !is Input.Empty -> true
+        formState.carbohydrates !is Input.Empty -> true
+        formState.fats !is Input.Empty -> true
+        formState.sugars !is Input.Empty -> true
+        formState.saturatedFats !is Input.Empty -> true
+        formState.salt !is Input.Empty -> true
+        formState.sodium !is Input.Empty -> true
+        formState.fiber !is Input.Empty -> true
+        formState.packageWeight !is Input.Empty -> true
+        formState.servingWeight !is Input.Empty -> true
         else -> false
     }
 
-    private val _eventBus = Channel<ProductFormEvent>()
-    val eventBus = _eventBus.receiveAsFlow()
-
-    fun onUpdate() {
-        val formState = _formState.value ?: return
+    fun onCreateProduct() {
+        val formState = _formState.value
 
         if (formState.error != null || !formState.isValid) {
             return
@@ -319,18 +293,17 @@ internal class UpdateProductViewModel(
             formState.servingWeight.takeIf { it.isValidOrEmpty }?.value?.toFloatOrNull()
 
         viewModelScope.launch {
-            _eventBus.send(ProductFormEvent.UpdatingProduct)
+            _eventBus.send(ProductFormEvent.CreatingProduct)
 
-            productRepository.updateProduct(
-                id = productId,
+            val id = productRepository.createProduct(
                 name = name,
                 brand = brand,
                 barcode = barcode,
                 calories = calories,
                 proteins = proteins,
                 carbohydrates = carbohydrates,
-                sugars = sugars,
                 fats = fats,
+                sugars = sugars,
                 saturatedFats = saturatedFats,
                 salt = salt,
                 sodium = sodium,
@@ -339,7 +312,99 @@ internal class UpdateProductViewModel(
                 servingWeight = servingWeight
             )
 
-            _eventBus.send(ProductFormEvent.ProductUpdated(productId))
+            _eventBus.send(ProductFormEvent.ProductCreated(id))
         }
+    }
+
+    private val _openFoodFactsLink = MutableStateFlow(input<OpenFoodFactsLinkError>())
+    val openFoodFactsLink = _openFoodFactsLink.asStateFlow()
+    private val openFoodFactsLinkForm = Form<OpenFoodFactsLinkError>(
+        strategy = ValidationStrategy.FailFast,
+        OpenFoodFactsLinkRules.NotEmpty,
+        OpenFoodFactsLinkRules.ValidUrl
+    )
+
+    fun onDownloadLinkChange(link: String) {
+        _openFoodFactsLink.update {
+            openFoodFactsLinkForm.validate(link)
+        }
+    }
+
+    private val openFoodFactsLinkHelper by lazy { OpenFoodFactsLinkHelper() }
+    private val _openFoodFactsError = MutableStateFlow<DownloadProductFailed?>(null)
+    val openFoodFactsError = _openFoodFactsError.asStateFlow()
+    fun onDownloadOpenFoodFacts() {
+        val url = _openFoodFactsLink.value.value
+
+        if (url.isEmpty()) {
+            _openFoodFactsLink.update {
+                Input.Invalid(it.value, listOf(OpenFoodFactsLinkError.Empty))
+            }
+            return
+        }
+
+        viewModelScope.launch {
+            _openFoodFactsError.emit(null)
+
+            val code = when (val code = openFoodFactsLinkHelper.extractCode(url)) {
+                null -> {
+                    _openFoodFactsLink.update {
+                        Input.Invalid(it.value, listOf(OpenFoodFactsLinkError.InvalidUrl))
+                    }
+                    return@launch
+                }
+
+                else -> code
+            }
+
+            _isDownloading.emit(true)
+
+            val product = runCatching {
+                openFoodFactsRemoteDataSource.getProduct(code, null) ?: error("Product not found")
+            }
+
+            product
+                .onSuccess { remote ->
+                    _formState.update {
+                        val newState = remote.asState()
+                        newState.copy(
+                            isModified = isModified(newState)
+                        )
+                    }
+                    _eventBus.send(ProductFormEvent.DownloadedProductSuccessfully)
+                    _openFoodFactsLink.update { input() }
+                    _openFoodFactsError.emit(null)
+                }
+                .onFailure {
+                    _openFoodFactsError.emit(DownloadProductFailed(it))
+                }
+
+            _isDownloading.emit(false)
+        }
+    }
+
+    private fun OpenFoodFactsProduct.asState(): ProductFormState = ProductFormState(
+        name = nameForm(productName ?: ""),
+        brand = brandForm(brands ?: ""),
+        barcode = barcodeForm(code ?: ""),
+        proteins = proteinsForm(nutrients?.proteins100g?.formatClipZeros() ?: ""),
+        carbohydrates = carbohydratesForm(nutrients?.carbohydrates100g?.formatClipZeros() ?: ""),
+        fats = fatsForm(nutrients?.fat100g?.formatClipZeros() ?: ""),
+        sugars = sugarsForm(nutrients?.sugars100g?.formatClipZeros() ?: ""),
+        saturatedFats = saturatedFatsForm(nutrients?.saturatedFat100g?.formatClipZeros() ?: ""),
+        salt = saltForm(nutrients?.salt100g?.formatClipZeros() ?: ""),
+        sodium = sodiumForm(nutrients?.sodium100g?.formatClipZeros() ?: ""),
+        fiber = fiberForm(nutrients?.fiber100g?.formatClipZeros() ?: ""),
+        packageWeight = packageWeightForm(packageQuantity?.formatClipZeros() ?: ""),
+        servingWeight = servingWeightForm(servingQuantity?.formatClipZeros() ?: "")
+    )
+}
+
+private class OpenFoodFactsLinkHelper {
+    fun extractCode(url: String): String? {
+        // Extract barcode from product URL
+        val regex = "openfoodfacts\\.org/product/(\\d+)".toRegex()
+        val barcode = regex.find(url)?.groupValues?.getOrNull(1)
+        return barcode
     }
 }
