+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/infrastructure/di/InitKoin.kt
@@ -9,7 +9,7 @@ import com.maksimowiczm.foodyou.feature.goals.goalsModule
 import com.maksimowiczm.foodyou.feature.language.languageModule
 import com.maksimowiczm.foodyou.feature.meal.mealModule
 import com.maksimowiczm.foodyou.feature.measurement.measurementModule
-import com.maksimowiczm.foodyou.feature.productredesign.productModule
+import com.maksimowiczm.foodyou.feature.product.productModule
 import com.maksimowiczm.foodyou.feature.recipe.recipeModule
 import org.koin.core.context.startKoin
 import org.koin.dsl.KoinAppDeclaration
